refactor: location service
