+++ lib/route/settings/content/root.dart
@@ -1,9 +1,9 @@
 import 'dart:io';
 
 import 'package:clipboard/clipboard.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:dpip/core/notify.dart';
 import 'package:dpip/widget/list/tile_group_header.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:material_symbols_icons/symbols.dart';
 
@@ -83,9 +83,7 @@ class _SettingsRootViewState extends State<SettingsRootView> {
           const ListTileGroupHeader(title: "複製 FCM Token"),
           ListTile(
             leading: Icon(
-              Platform.isAndroid
-                ? Icons.bug_report_rounded
-                : CupertinoIcons.square_on_square,
+              Platform.isAndroid ? Icons.bug_report_rounded : CupertinoIcons.square_on_square,
             ),
             title: const Text("複製 FCM Token"),
             onTap: () {
